+++ src/sound/SoundDriver.hh
@@ -1,8 +1,6 @@
 #ifndef SOUNDDRIVER_HH
 #define SOUNDDRIVER_HH
 
-#include <cstdint>
-
 namespace openmsx {
 
 class SoundDriver
@@ -28,7 +26,7 @@ public:
 	  */
 	virtual unsigned getSamples() const = 0;
 
-	virtual void uploadBuffer(int16_t* buffer, unsigned len) = 0;
+	virtual void uploadBuffer(float* buffer, unsigned len) = 0;
 
 protected:
 	SoundDriver() = default;
